main/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
dateparser==1.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-debug-toolbar==4.3.0
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.3
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pybit==5.6.2
pycparser==2.21
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
redis==5.0.4
celery==5.4.0
